centaurus

+++ SA/state.h
@@ -11,7 +11,7 @@ typedef struct
 
 int _pop_ideal(int d);
 void popul_test1(TU **units, int n, int k, int ideal_pop);
-long long energy_population(TU **units, Cluster *cluster, int margin, int k, int n, int ideal_pop);
+int energy_population(TU **units, Cluster *cluster, int margin, int k, int n, int ideal_pop);
 int energy_compactness(Cluster *clusters, int k);
 /* The energy of the current state */
 int
